RL - init autoregression

+++ ETA/model.py
@@ -163,7 +163,7 @@ class Model(tf_keras.Model):
         tf.summary.scalar(
             name="Q/assert",
             data=self.ttr_param,
-            step=self.counter,
+            step=tf.cast(self.counter, tf.int64),
         )
         tf.summary.scalar(
             name="Q/table",
